Adding main pages
