+++ src/views/InstructionView.vue
@@ -1,0 +1,23 @@
+<template>
+    <div class="about">
+      <v-card>
+        <v-card-title> This is a vue/vuetify class! </v-card-title>
+        <v-card-subtitle> (College Class) </v-card-subtitle>
+  
+        <v-card-text>
+          Lorem ipsum dolor sit amet blah blah blah
+          <v-icon icon="mdi:mdi-abacus" />
+        </v-card-text>
+      </v-card>
+    </div>
+  </template>
+  
+  <style>
+  @media (min-width: 1024px) {
+    .about {
+      min-height: 100vh;
+      display: flex;
+      align-items: center;
+    }
+  }
+  </style>
